修改玩家 id 类型

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/GameVirtualOrder.java
@@ -41,7 +41,7 @@ public class GameVirtualOrder implements Serializable {
      * 玩家id
      */
     @Excel(name = "玩家id", width = 15)
-    private java.lang.Integer playerId;
+    private java.lang.Long playerId;
 
     /**
      * 商品id
